package creating + rename builting loader + compress style into a single file
