--- utils/style/constants.rb
@@ -1,6 +1,0 @@
-module Style
-  ESCAPE_CHARACTERS = {
-    '&' => '&&',
-    '<' => '&'
-  }.freeze
-end
